+++ src/app/core/models/channel.ts
