+++ src/app/core/models/file.ts
