+++ src/app/core/models/message.ts
