+++ src/app/main/side-nav/apps/apps.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { SideNavService } from '../../../core/side-nav.service';
+import { SideNavService } from 'src/app/core/side-nav.service';
 
 @Component({
   selector: 'app-apps',
