+++ src/app/main/side-nav/direct-messages/direct-messages.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ChatService } from 'src/app/core/chat.service';
-import { DirectMessage } from 'src/app/core/direct-message';
-import { User } from 'src/app/core/user';
-import { SideNavService } from '../../../core/side-nav.service';
+import { DirectMessage } from 'src/app/core/models/direct-message';
+import { User } from 'src/app/core/models/user';
+import { SideNavService } from 'src/app/core/side-nav.service';
 
 @Component({
   selector: 'app-direct-messages',
